?????? ??????? 1839 Change int32 type to float in ItemStatValue. ????...

+++ src/game/ItemPrototype.h
@@ -452,8 +452,8 @@ struct _Damage
 struct _ItemStat
 {
     uint32  ItemStatType;
-    //int32   ItemStatValue;
-	float	ItemStatValue;
+    int32   ItemStatValue;
+	//float	ItemStatValue;
 };
 struct _Spell
 {
